# Fix perf of multiline semantic tokens, and LF handling of same