--- a/src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/RazorSemanticTokenInfoServiceTest/GetSemanticTokens_Razor_MultiLineCommentWithBlankLines_LF.semantic.txt
+++ b/src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Semantic/TestFiles/RazorSemanticTokenInfoServiceTest/GetSemanticTokens_Razor_MultiLineCommentWithBlankLines_LF.semantic.txt
@@ -1,0 +1,10 @@
+//line,characterPos,length,tokenType,modifier
+0 0 1 razorCommentTransition 0
+0 1 1 razorCommentStar 0
+0 1 4 razorComment 0
+2 0 3 razorComment 0
+1 0 4 razorComment 0
+1 0 19 razorComment 0
+1 0 3 razorComment 0
+0 3 1 razorCommentStar 0
+0 1 1 razorCommentTransition 0