Fixes #6756
Fixes #8176

The compiler team fixed the perf of `GetContent()` in https://github.com/dotnet/razor/pull/8032 but we were still being pretty naive in our usage here, and the use of `Environment.NewLine` was just wrong.

Results show no harm without a large comment, but a big saving in allocations when there is a large comment.

Old:

|                        Method | WithMultiLineComment |     Mean |    Error |   StdDev |   Median |      Min |       Max |   Gen0 |   Gen1 | Allocated |
|------------------------------ |--------------------- |---------:|---------:|---------:|---------:|---------:|----------:|-------:|-------:|----------:|
| RazorSemanticTokensRangeAsync |                False | 53.43 us | 1.057 us | 2.428 us | 53.54 us | 48.58 us |  60.56 us | 0.6104 |      - |   3.75 KB |
| RazorSemanticTokensRangeAsync |                 True | 97.77 us | 1.787 us | 1.912 us | 98.18 us | 93.45 us | 100.36 us | 6.1035 | 0.2441 |  34.87 KB |

New:

|                        Method | WithMultiLineComment |     Mean |    Error |   StdDev |   Median |      Min |       Max |   Gen0 |   Gen1 | Allocated |
|------------------------------ |--------------------- |---------:|---------:|---------:|---------:|---------:|----------:|-------:|-------:|----------:|
| RazorSemanticTokensRangeAsync |                False | 53.50 us | 1.023 us | 1.218 us | 53.63 us | 50.81 us |  55.36 us | 0.6104 | 0.0610 |   3.75 KB |
| RazorSemanticTokensRangeAsync |                 True | 83.66 us | 1.581 us | 1.401 us | 83.25 us | 82.27 us |  86.72 us | 2.6855 | 0.1221 |  15.29 KB |